requirements.txt
Flask
gunicorn
psycopg2-binary
Flask-SQLAlchemy
pyserial
paho-mqtt
python-dotenv
Flask-CORS 